requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
gunicorn==19.9.0
Werkzeug==1.0.1
Jinja2==2.11.2
newsapi-python==0.2.6
MarkupSafe==1.1.1
itsdangerous==1.1.0
click==7.1.2